requirements.txt
python-dotenv==1.0.1

requests==2.32.3

slack-bolt==1.19.0

slack-sdk==3.33.1

Flask==3.0.3

schedule==1.2.2

croniter==6.0.0

matplotlib==3.9.2

pandas==2.2.2

numpy==2.1.0

SQLAlchemy==2.0.36

psycopg2-binary==2.9.9

langchain==0.2.16

langchain-community==0.2.16

langchain-openai==0.1.22

openai==1.44.0

faiss-cpu==1.8.0.post1

sqlglot==25.20.2

mindsdb-sdk==1.2.0

pyslop==2.3.0